requirements.txt
python-dotenv
fastapi
uvicorn
ruff
dependency-injector